backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
requests==2.32.5
numpy==2.2.6
pandas==2.3.2
scikit-learn==1.7.1
joblib==1.5.1
shap==0.48.0
xgboost==3.0.4
yfinance==0.2.65
